feat: use const fn where possible
this just enables the use of the function inside other const contexts.
one potential caveat to consider is if you ever need to make this
method non-const in future, it's a breaking change.
that seems unlikely in this case, but it's worth bearing in mind.

suggested by clippy's `missing_const_for_fn` lint:
https://rust-lang.github.io/rust-clippy/master/index.html#/missing_const_for_fn